requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.4
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx2pdf==0.1.8
drf-yasg==1.21.10
fire==0.7.0
flower==2.0.1
fonttools==4.58.5
gunicorn==22.0.0
humanize==4.12.3
inflection==0.5.1
kombu==5.5.4
lxml==6.0.0
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==25.0
pdf2docx==0.5.8
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMuPDF==1.26.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
termcolor==3.1.0
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
